requirements.txt
apache-airflow==2.9.1
apache-airflow-providers-docker